requirements.txt
flask
gunicorn==19.9.0
line-bot-sdk
pillow